fix(main.ts): fix typing for .vue files
declare a TypeScript module for every file ending with *.vue (SFC). TypeScript will digest such a
module as having the following: An import statement for the defineComponent global method,
Declaration of a 'component' variable of type 'defineComponent', a default export of the 'component'
variable. Having the shim file above allows TypeScript to consider any Vue Component to be a
well-defined module